Add files via upload
Today I worked on creating a new python file that could take in different images and resize them, while also being able to identify them. I also was able to set the boundaries for a new hut that players could interact with and finally am working on a new layout for the second map of my game.