Add files via upload
Today I worked on creating the parameters of my second world. I also tried to fix the lag of my second world but it didn’t work, so instead I just changed the player speed. I also spend a majority of the time trying to think of what I would do with this next world.